+++ frontend/src/Ask&Answer/AskAndAnswer.jsx
@@ -14,7 +14,6 @@ const Questions = () => {
             <h5>4 questions</h5>
             <Link variant="primary" to="/createquestion" className="mb-2">Ask  Question</Link>
             <Link variant="primary" to="/myquestions" className="mb-2">My  Questions</Link>
-
             <AllQA/>
            
         </Container>
